move to seperate file and use an in context flag
