+++ util.go
@@ -26,6 +26,7 @@ const (
 	arrowBatches          contextKey = "ARROW_BATCHES"
 	queryTag              contextKey = "QUERY_TAG"
 	submitSync            contextKey = "SUBMIT_SYNC"
+	reportAsyncError      contextKey = "REPORT_ASYNC_ERROR"
 )
 
 const (
@@ -108,6 +109,12 @@ func WithQueryTag(ctx context.Context, tag string) context.Context {
 // can poll for status using the query ID.
 func WithSubmitSync(ctx context.Context) context.Context {
 	return context.WithValue(ctx, submitSync, true)
+}
+
+// WithReportAsyncError returns a context that enables execution to panic and return
+// any data that could be useful for debugging waitForCompletedQueryResultResp
+func WithReportAsyncError(ctx context.Context) context.Context {
+	return context.WithValue(ctx, reportAsyncError, true)
 }
 
 // Get the request ID from the context if specified, otherwise generate one
